refactor(home): extract model-state plumbing into useHomeModelState
